chore: relocate completion scripts
