+++ packages/cli/src/completions/index.ts
