+++ packages/cli/src/completions/phantom-fish.ts
