+++ packages/cli/src/completions/phantom-zsh.ts
