+++ packages/cli/src/completions/phantom.bash
