+++ packages/cli/src/completions/phantom.zsh
